expression: move some funcs out of builtin.go
