+++ expression/builtin_common.go
@@ -1,0 +1,239 @@
+// Copyright 2017 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package expression
+
+import (
+	"sort"
+
+	"github.com/juju/errors"
+	"github.com/pingcap/tidb/context"
+	"github.com/pingcap/tidb/util/types"
+)
+
+var (
+	_ functionClass = &coalesceFunctionClass{}
+	_ functionClass = &isNullFunctionClass{}
+	_ functionClass = &greatestFunctionClass{}
+	_ functionClass = &leastFunctionClass{}
+	_ functionClass = &intervalFunctionClass{}
+)
+
+var (
+	_ builtinFunc = &builtinCoalesceSig{}
+	_ builtinFunc = &builtinIsNullSig{}
+	_ builtinFunc = &builtinGreatestSig{}
+	_ builtinFunc = &builtinLeastSig{}
+	_ builtinFunc = &builtinIntervalSig{}
+)
+
+type coalesceFunctionClass struct {
+	baseFunctionClass
+}
+
+func (c *coalesceFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
+	return &builtinCoalesceSig{newBaseBuiltinFunc(args, ctx)}, errors.Trace(c.verifyArgs(args))
+}
+
+type builtinCoalesceSig struct {
+	baseBuiltinFunc
+}
+
+func (b *builtinCoalesceSig) eval(row []types.Datum) (types.Datum, error) {
+	args, err := b.evalArgs(row)
+	if err != nil {
+		return types.Datum{}, errors.Trace(err)
+	}
+	return builtinCoalesce(args, b.ctx)
+}
+
+// See http://dev.mysql.com/doc/refman/5.7/en/comparison-operators.html#function_coalesce
+func builtinCoalesce(args []types.Datum, ctx context.Context) (d types.Datum, err error) {
+	for _, d = range args {
+		if !d.IsNull() {
+			return d, nil
+		}
+	}
+	return d, nil
+}
+
+type isNullFunctionClass struct {
+	baseFunctionClass
+}
+
+func (c *isNullFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
+	return &builtinIsNullSig{newBaseBuiltinFunc(args, ctx)}, errors.Trace(c.verifyArgs(args))
+}
+
+type builtinIsNullSig struct {
+	baseBuiltinFunc
+}
+
+func (b *builtinIsNullSig) eval(row []types.Datum) (types.Datum, error) {
+	args, err := b.evalArgs(row)
+	if err != nil {
+		return types.Datum{}, errors.Trace(err)
+	}
+	return builtinIsNull(args, b.ctx)
+}
+
+// See https://dev.mysql.com/doc/refman/5.7/en/comparison-operators.html#function_isnull
+func builtinIsNull(args []types.Datum, _ context.Context) (d types.Datum, err error) {
+	if args[0].IsNull() {
+		d.SetInt64(1)
+	} else {
+		d.SetInt64(0)
+	}
+	return d, nil
+}
+
+type greatestFunctionClass struct {
+	baseFunctionClass
+}
+
+func (c *greatestFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
+	return &builtinGreatestSig{newBaseBuiltinFunc(args, ctx)}, errors.Trace(c.verifyArgs(args))
+}
+
+type builtinGreatestSig struct {
+	baseBuiltinFunc
+}
+
+func (b *builtinGreatestSig) eval(row []types.Datum) (types.Datum, error) {
+	args, err := b.evalArgs(row)
+	if err != nil {
+		return types.Datum{}, errors.Trace(err)
+	}
+	return builtinGreatest(args, b.ctx)
+}
+
+// See http://dev.mysql.com/doc/refman/5.7/en/comparison-operators.html#function_greatest
+func builtinGreatest(args []types.Datum, ctx context.Context) (d types.Datum, err error) {
+	if args[0].IsNull() {
+		return
+	}
+	max := 0
+	sc := ctx.GetSessionVars().StmtCtx
+	for i := 1; i < len(args); i++ {
+		if args[i].IsNull() {
+			return
+		}
+
+		var cmp int
+		if cmp, err = args[i].CompareDatum(sc, args[max]); err != nil {
+			return
+		}
+
+		if cmp > 0 {
+			max = i
+		}
+	}
+	d = args[max]
+	return
+}
+
+type leastFunctionClass struct {
+	baseFunctionClass
+}
+
+func (c *leastFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
+	return &builtinLeastSig{newBaseBuiltinFunc(args, ctx)}, errors.Trace(c.verifyArgs(args))
+}
+
+type builtinLeastSig struct {
+	baseBuiltinFunc
+}
+
+func (b *builtinLeastSig) eval(row []types.Datum) (types.Datum, error) {
+	args, err := b.evalArgs(row)
+	if err != nil {
+		return types.Datum{}, errors.Trace(err)
+	}
+	return builtinLeast(args, b.ctx)
+}
+
+// See http://dev.mysql.com/doc/refman/5.7/en/comparison-operators.html#function_least
+func builtinLeast(args []types.Datum, ctx context.Context) (d types.Datum, err error) {
+	if args[0].IsNull() {
+		return
+	}
+	min := 0
+	sc := ctx.GetSessionVars().StmtCtx
+	for i := 1; i < len(args); i++ {
+		if args[i].IsNull() {
+			return
+		}
+
+		var cmp int
+		if cmp, err = args[i].CompareDatum(sc, args[min]); err != nil {
+			return
+		}
+
+		if cmp < 0 {
+			min = i
+		}
+	}
+	d = args[min]
+	return
+}
+
+type intervalFunctionClass struct {
+	baseFunctionClass
+}
+
+func (c *intervalFunctionClass) getFunction(args []Expression, ctx context.Context) (builtinFunc, error) {
+	return &builtinIntervalSig{newBaseBuiltinFunc(args, ctx)}, errors.Trace(c.verifyArgs(args))
+}
+
+type builtinIntervalSig struct {
+	baseBuiltinFunc
+}
+
+func (b *builtinIntervalSig) eval(row []types.Datum) (types.Datum, error) {
+	args, err := b.evalArgs(row)
+	if err != nil {
+		return types.Datum{}, errors.Trace(err)
+	}
+	return builtinInterval(args, b.ctx)
+}
+
+// See http://dev.mysql.com/doc/refman/5.7/en/comparison-operators.html#function_interval
+func builtinInterval(args []types.Datum, ctx context.Context) (d types.Datum, err error) {
+	if args[0].IsNull() {
+		d.SetInt64(int64(-1))
+		return
+	}
+	sc := ctx.GetSessionVars().StmtCtx
+
+	idx := sort.Search(len(args)-1, func(i int) bool {
+		d1, d2 := args[0], args[i+1]
+		if d1.Kind() == types.KindInt64 && d1.Kind() == d2.Kind() {
+			return d1.GetInt64() < d2.GetInt64()
+		}
+		if d1.Kind() == types.KindUint64 && d1.Kind() == d2.Kind() {
+			return d1.GetUint64() < d2.GetUint64()
+		}
+		if d1.Kind() == types.KindInt64 && d2.Kind() == types.KindUint64 {
+			return d1.GetInt64() < 0 || d1.GetUint64() < d2.GetUint64()
+		}
+		if d1.Kind() == types.KindUint64 && d2.Kind() == types.KindInt64 {
+			return d2.GetInt64() > 0 && d1.GetUint64() < d2.GetUint64()
+		}
+		v1, _ := d1.ToFloat64(sc)
+		v2, _ := d2.ToFloat64(sc)
+		return v1 < v2
+	})
+	d.SetInt64(int64(idx))
+
+	return
+}
